+++ src/daoFactory.js
@@ -1,0 +1,5 @@
+const jobDao = require('./daos/job');
+
+module.exports = {
+  job: new jobDao(),
+};
